* master:
  Rnamed dequeue to pop and enqueue to push
  Added RingBuffer implementation
  Changed exo message structure order. Moved exo_magic to beginning of header so that exo messages are easier to find the true start of in a seekable stream. Added feature to enable the explicit passing of an open file descriptor to Pipeline Inlet and Outlet objects.
  Added check in log Statement destructor to prevent segfault if logger is null
  Ignore multi lined comments, fixed gcc bug
  Fixed errors uncovered by strict compilation errors
  Small fix, increasing compiler strickness for tests
  Fixed small bug in for_each_path
  Added SO_REUSEPORT for inlet ports

+++ inc/nix/config.hpp
@@ -209,7 +209,7 @@ struct Config
             if (fd == -1) { return exo::Result::DOES_NOT_EXIST; }
 
             auto size = sizeof(blob);
-            if (read(fd, &blob, size) != size)
+            if (read(fd, &blob, size) != (int)size)
             {
                 close(fd);
                 return exo::Result::READ_ERR;
@@ -233,7 +233,7 @@ struct Config
             if (fd > -1)
             {
                 auto size = sizeof(blob);
-                if (write(fd, &blob, size) != size)
+                if (write(fd, &blob, size) != (int)size)
                 {
                     close(fd);
                     return exo::Result::WRITE_ERR;
